Emphasize concept sketch state; Add human facilitators, Estimates, Thesis, business case
Pylon teams will only be as useful as they let us actually operate a business. Add some hints as to what this might look like in practice to try to pique people's interest, and a few concrete types (Estimates, Thesis) that I already feel confident that we need (Estimates are for dashboards; Thesis is for important wiki pages, definitely norms and process, ideally "how it works" though I'm still skeptical about that case).

My first few attempts at vibe-coding have failed spectacularly — it's no surprise that people are getting pwned by it! The coding agents still don't really understand what they're doing, and they consistently make mistakes which are both subtle and spectacularly dumb, arising from complete lack of understanding of where we're going next, which takes us completely off the rails and closes the path to future feature development in ways that are subtle, pervasive, deeply engrained, and hard to understand/troubleshoot! (I see them crashing out first by overcomplicating, a classic problem to which humans are also vulnerable, and frequently impossible to unwind!). 

So next up I'm going to try writing a pile of architecture docs, with the goal of building a big long context with lots of instructions about how to approach coding in order to constrain the LLMs' search space. It's really wild how creative they get at making mistakes or doing dumb shit that prevents upcoming features! I will probably commit these architecture docs, with additional caveats on the README, to try to start nerd-sniping software engineers, and we'll see if they help.

The human facilitators element, I think, is actually the point of Pylon. A wise man once said, "organizational software is the skin on the sausage" (where the meat of the sausage, the muscle, is actual human relationships and working practices) — the point of AI is to augment human intelligence. So Pylon is really skills / effectiveness augmentation for post-capitalist facilitators — an AI agent to "do their bidding" in facilitating the group.